Add comma as an alternative options string separator (#909)
Add comma as an alternative options string separator, in addition to
whitespace.

This is useful for multi-level nested bash scripts. Google has some of
these for building GoogleJDK or running benchmarks. Using comma
separators can simplify the script a lot, and can keep options as a
single argument without dealing with any string or whitespace escaping.

---------

Co-authored-by: Zixian Cai <2891235+caizixian@users.noreply.github.com>